requirements.txt
accelerate==0.22.0
av==14.0.0
bitsandbytes==0.41.1
deepspeed==0.11.1
einops==0.4.1
fastapi==0.100.1
ftfy==6.3.1
fvcore==0.1.5.post20221221
gradio==3.39.0
ninja==1.11.1.1
numpy==1.24.2
openai==0.27.8
opencv-python==4.8.0.74
packaging
peft==0.4.0
Pillow==9.4.0
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
ray==2.6.1
safetensors==0.3.3
scipy==1.11.2
sentencepiece==0.1.99
shortuuid==1.0.11
timm==0.5.4
tokenizers==0.13.3
torch==1.13.1+cu117
torchaudio==0.13.0
torchvision==0.14.1+cu117
tqdm==4.64.1
transformers==4.31.0
typing_extensions==4.7.1
wandb==0.15.12